requirements.txt
uwsgi
bcrypt
Flask
Flask-Login
Flask-WTF
mysql-connector-python
